+++ visualization/app/codeCharta/features/codeMap/codeMap.render.service.ts
@@ -7,6 +7,7 @@ import { labelsPerMapActiveSelector, nodeMetricDataSelector } from "../../render
 import {
     CodeMapMesh,
     ColorCategoryCountsStore,
+    createAreaTrueTreemapNodes,
     createTreemapNodes,
     StreetLayoutGenerator,
     ThreeSceneService,
@@ -111,6 +112,8 @@ export class CodeMapRenderService implements OnDestroy, RendererEngine {
                 )
             case LayoutAlgorithm.SquarifiedTreeMap:
                 return createTreemapNodes(map, state, nodeMetricData, deltaState)
+            case LayoutAlgorithm.AreaTrueTreemap:
+                return createAreaTrueTreemapNodes(map, state, nodeMetricData, deltaState)
             default:
                 return []
         }